1、优化代码结构 2、初步完成多任务下载功能，需要优化

+++ app/src/main/java/xpfei/myapp/view/progresslayout/ProgressLayoutListener.java
@@ -18,5 +18,5 @@ package xpfei.myapp.view.progresslayout;
 public interface ProgressLayoutListener {
     void onProgressCompleted();
 
-    void onProgressChanged(int seconds);
+    void onProgressChanged(long seconds);
 }
